Implement know-merge
know-merge merges several graphs (as opposed to taking the union, which
could be done with `cat`). The current implementation parses all files
into a graph and then serializes the result. This is clearly suboptimal.
It would be better to parse triple by triple and write it after
renaming of blank nodes where necessary.